# Performance tweaks to SpanAdapter