- Don't allocate `KeyValue.Builder` when not used
- Use singleton `Status` proto if no description
- `computeIfAbsent` / forEach because Java 8 now
- Cache ID bytes conversion (we'll expect many cache hits for trace ID, and some cache hits for span ID which are parents of other spans)

After
```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.803 ±   0.013   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1692.960 ±   8.383  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  7472.003 ±  12.027    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1694.252 ±  43.026  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  7477.808 ± 187.806    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.004 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.017 ±   0.004    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   462.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   211.000                ms
```

Before
```
Benchmark                                                  Mode  Cnt      Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45      3.459 ±   0.143   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45   1862.199 ±  53.989  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  10101.337 ±  82.186    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45   1866.701 ±  66.144  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  10122.074 ± 193.132    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45      0.006 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45      0.033 ±   0.007    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45    551.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45    238.000                ms
```