@anuraaga 2021-06-04
Combined with #3285 ends up as

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.921 ±   0.042   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1699.122 ±  23.175  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  7813.337 ±  57.820    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1704.615 ±  45.854  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  7837.734 ± 183.193    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.004 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.020 ±   0.004    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   494.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   216.000                ms
```