@anuraaga 2021-06-04
Escape analysis doesn't help enough, `HashMap.forEach` saves a reasonable amount

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     3.361 ±   0.017   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1841.403 ±   8.465  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  9744.004 ±  24.055    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1844.216 ±  44.524  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  9758.406 ± 226.496    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.005 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.026 ±   0.006    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   503.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   227.000                ms
```