@anuraaga 2021-06-04
Cache ID conversion

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.843 ±   0.012   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  2009.912 ±  11.773  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  9000.003 ±  24.055    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  2010.395 ±  36.452  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  9002.223 ± 157.470    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.005 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.024 ±   0.005    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   583.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   258.000                ms
```