@anuraaga 2021-06-04
Don't allocate separate list for IL protos (it's copied into yet another list anyways)

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.877 ±   0.141   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1966.587 ±  73.000  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  8853.337 ±  22.322    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1972.137 ±  74.405  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  8882.796 ± 149.932    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.006 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.025 ±   0.006    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   605.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   279.000                ms
```