@anuraaga 2021-06-04
Ok couldn't resist. Proto generated code is large so I must be missing something, but 1KB+ savings is too much to resist

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.803 ±   0.013   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1692.960 ±   8.383  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  7472.003 ±  12.027    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1694.252 ±  43.026  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  7477.808 ± 187.806    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.004 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.017 ±   0.004    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   462.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   211.000                ms
```